SSH Scripts Functioning
can add/delete/update servers
can add/delete/update scripts
can add/delete scripts to/from servers
can set interval per server and default for scripts
included navbar everywhere
specified angular version in bower.json
moved script management logic to server api controller
added authentication to server and script management pages
switched client/server tests in grunt...they were crashing phantomjs
added asyncjs for easier code management
changed base scripts so they only return a number
rest api interval starter, probably should only be for tests
created some basic integration tests for the server rest api
added ssh key environment variable
grunt is still having problems in spite of my efforts

+++ client/app/server/server.controller.js
@@ -1,24 +1,52 @@
 'use strict';
 
 angular.module('moreOnApp')
-  .controller('ServerCtrl', function ($scope, $http, $routeParams) {
-    $http.get('/api/v1/servers/' + $routeParams.id)
+  .controller('ServerCtrl', function ($scope, $http, $routeParams, Modal, $window) {
+  	$scope.server = {};
+  	$scope.scripts = [];
+  	$scope.alerts = [];	
+
+  	function GetServer () {
+  		$http.get('/api/v1/servers/' + $routeParams.id)
+		    .success(function (data) {
+		    	$scope.server = data;
+		    });
+  	}
+
+  	new GetServer();
+
+	  $http.get('/api/v1/scripts')
 	    .success(function (data) {
-	    	$scope.server = data;
+	    	$scope.scripts = data;
 	    });
+
+	  $scope.addScript = function (script) {
+	  	$scope.server.activeScripts.push({
+	  		script: script,
+	  		duration: script.defaultInterval,
+	  		server: $scope.server._id
+	  	});
+	  };
 
 	  $scope.update = function () {
 	  	$http.put('/api/v1/servers/' + $routeParams.id, $scope.server)
-	  		.success(function (data) {
-	  			// TODO: notify user
+	  		.success(function () {
+	  			$scope.alerts.push({
+	  				type: 'success',
+	  				message: 'Save Successful!'
+	  			});
 	  		});
 	  };
 
 	  $scope.delete = function () {
-	  	$http.delete('/api/v1/servers/' + $routeParams.id)
-	  		.success(function (data) {
-	  			// TODO: redirect to servers page
-	  		});
+	  	var deleteModal = Modal.confirm.delete(function () {
+	  		$http.delete('/api/v1/servers/' + $routeParams.id)
+		  		.success(function () {
+		  			$window.location.href = '/server';
+		  		});
+	  	});
+
+	  	deleteModal($scope.server.hostname);
 	  };
   });
